Move pages and tests to app name directory
